Research_Paper_Data_Pipeline/requirements.txt
# Python dependencies
requests>=2.25.0
pymongo>=4.0.0
feedparser>=6.0.0
python-dotenv>=0.20.0
schedule>=1.1.0

# API Clients
arxiv==1.4.3
crossref-commons==0.0.7
scholarly==1.7.11
semanticscholar==0.4.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.2

# Database
pymongo==4.4.1
motor==3.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.0.3
loguru==0.7.0
schedule==1.2.0
pytz==2023.3
tenacity==8.2.2
backoff==2.2.1

# Testing
pytest==7.3.1
pytest-mock==3.10.0
pytest-asyncio==0.21.1